chore(signup): code refactor
