+++ packages/blockchain-wallet-v4-frontend/src/scenes/Signup/types.ts
@@ -8,13 +8,7 @@ export type SignupFormInitValuesType = {
   country?: string
   email?: string
 }
-export type GeoLocationType = {
-  countryCode: string
-  headerBlockchainCFIpCountry: string
-  headerBlockchainGoogleIpCountry: string
-  headerBlockchainGoogleIpRegion: string
-  ip: string
-}
+
 export type SignupFormType = {
   country: string
   email: string
